🔨 Refactor : SearchPlace 리팩토링
+ SearchPlaceRefactoring
