+++ src/components/common/card/PlaceDetailCard.tsx
@@ -210,12 +210,6 @@ let ButtonGroup = {
   justifyContent: "space-between",
 }
 
-export interface ListDetailCardProp {
-  index: number
-  titleTop: string
-  titleBody: string
-  titleBottom: string
-}
 interface CoursePlace extends Place {
   description: string
 }
